Mock get function for contacts
