--- routes/addon.routes.js
@@ -1,6 +1,0 @@
-let routes = require('express').Router()
-let controller = require('../controllers/addon.controller')
-
-routes.get('/:email', controller.get)
-
-module.exports = routes
